// src/app/shared/generated/model/eximBookmarkScope.ts
/**
 * onecx-bookmark-bff
 * Backend-For-Frontend (BFF) service for onecx-bookmark. This API provides endpoints to manage Bookmarks
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum EximBookmarkScope {
    Private = 'PRIVATE',
    Public = 'PUBLIC'
}