// src/app/shared/generated/model/exportBookmarksRequest.ts
/**
 * onecx-bookmark-bff
 * Backend-For-Frontend (BFF) service for onecx bookmark. With this API you can manage bookmarks in your portal.
 *
 * Contact: tkit_dev@1000kit.org
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EximBookmarkScope } from './eximBookmarkScope';


export interface ExportBookmarksRequest { 
    workspaceName?: string;
    scopes: Array<EximBookmarkScope>;
}